pyproject.toml
[project]
name = "sam3dbody-standalone"
version = "0.1.0"
description = "Standalone WebUI for SAM 3D Body + SAM3 person mask (Win/Linux)"
requires-python = "==3.11.*"
dependencies = [
    # Web backend
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.41.0",
    "python-multipart>=0.0.9",

    # Core numerics / image IO
    "numpy>=1.26,<2.0",
    "scipy>=1.12.0",
    "pillow>=12.0.0",
    "opencv-python>=4.10",
    "safetensors>=0.7.0",

    # Torch stack — cu128 wheels from download.pytorch.org (Win amd64 + Linux x86_64 + Linux aarch64)
    "torch==2.10.0+cu128",
    "torchvision==0.25.0+cu128",
    "torchaudio==2.10.0+cu128",
    # xformers は意図的に外している: DINOv3 の LinearW24 の sparsify24 呼び出しだけが依存元で、
    # その関数は SAM3DBody の通常推論では起動しないため (実測で pipeline が xformers 無しで
    # warm 1.56s で完走、スコア不変)。約 350MB の install サイズを節約。復活させたい時は
    # `uv pip install xformers>=0.0.34 --index-url https://download.pytorch.org/whl/cu128`

    # SAM3 (Meta Segment Anything 3 — 公式 `sam3` を `src/sam3dbody_app/core/sam3/` に vendor 済み)
    "einops>=0.8.0",
    "ftfy>=6.1.0",
    "regex>=2024.0.0",
    "iopath>=0.1.10",
    "timm>=1.0.24",
    "scikit-image>=0.22.0",
    "av>=14.0",
    "pycocotools>=2.0.11",
    "tqdm",
    "psutil",
    "huggingface-hub>=0.23.0",
    "transformers>=4.50.0",

    # decord は video フレーム読込の fast path だが、Linux aarch64 に
    # wheel が無く、かつ image-mode 推論では使われないため外す。
    # aarch64 で動画推論 (Phase 4) を走らせたい場合は手動で pyav/opencv にフォールバック予定。
    "decord>=0.6.0; sys_platform != 'linux' or platform_machine != 'aarch64'",
    "triton-windows; sys_platform == 'win32'",
    "triton; sys_platform != 'win32'",

    # SAM 3D Body + MHR
    "trimesh[easy]>=4.0",
    "pyrender>=0.1.45",
    "PyOpenGL>=3.1.0",
    "pytorch-lightning>=2.0",
    "hydra-core>=1.3.0",
    "yacs>=0.1.8",
    "roma>=1.5.0",
    "networkx>=3.2",
    "pandas>=2.0",
    "rich>=13.0",
    "loguru>=0.7.0",
    "joblib>=1.2.0",
    "optree>=0.9.0",
    "ffmpeg-python>=0.2",
    "braceexpand>=0.1.7",
    # DINOv3 backbone の logging モジュールが colored() を使うため直接依存化
    "termcolor>=2.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sam3dbody_app"]

# ---------------------------------------------------------------------------
# uv-specific configuration
# ---------------------------------------------------------------------------
[tool.uv]
environments = [
    "sys_platform == 'win32' and platform_machine == 'AMD64' and python_version == '3.11'",
    "sys_platform == 'linux' and platform_machine == 'x86_64' and python_version == '3.11'",
    "sys_platform == 'linux' and platform_machine == 'aarch64' and python_version == '3.11'",
]

[tool.uv.sources]
torch        = [{ index = "pytorch-cu128" }]
torchvision  = [{ index = "pytorch-cu128" }]
torchaudio   = [{ index = "pytorch-cu128" }]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
